crown system (and formatting)
essentially, I implemented a forced zstack. i created a new div for the crown, set it to an absolute location and set the z-index to 0. then for the actual text, i set a z-index of 1. therefore, the stack will look like (bottom -> up) bubble (from the h1), crown (may be hidden), text. Adjust the crown properties under the css for the crown class.

and i did ⌥⌃F on a bunch of files